backend/requirements.txt
# Web Framework
Flask==3.0.0
flask-cors==4.0.0

# HTTP Requests
requests==2.31.0

# Data Parsing
feedparser==6.0.11
PyYAML==6.0.1

# Social Media APIs
praw==7.7.1            # Reddit API
tweepy==4.14.0         # Twitter API (if accessible)

# Google APIs
google-auth==2.25.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.110.0

# Background Tasks
APScheduler==3.10.4

# Environment Variables
python-dotenv==1.0.0

# Date/Time Utilities
python-dateutil==2.8.2

# Optional: Better logging
colorlog==6.8.0
